requirements.txt
requests==2.32.5
python-dotenv==1.2.2
anthropic>=0.40.0
openai>=1.0.0
